move Termination trait to std::process
