+++ src/libstd/lib.rs
@@ -502,10 +502,6 @@ mod memchr;
 // compiler
 pub mod rt;
 
-// The trait to support returning arbitrary types in the main function
-#[unstable(feature = "termination_trait", issue = "43301")]
-pub mod termination;
-
 // Include a number of private modules that exist solely to provide
 // the rustdoc documentation for primitive types. Using `include!`
 // because rustdoc only looks for these modules at the crate level.
